requirements.txt
asgiref==3.9.1
Brotli==1.1.0
cffi==1.17.1
charset-normalizer==3.4.3
cssselect2==0.8.0
Django==5.2.5
django-solo==2.4.0
fonttools==4.59.2
pdfkit==1.0.0
pillow==11.3.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
reportlab==4.4.3
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
weasyprint==66.0
webencodings==0.5.1
wkhtmltopdf==0.2
zopfli==0.2.3.post1
